client-python/requirements.txt
grpcio==1.65.4
grpcio-tools==1.65.4
protobuf
